# specs/001-define-nbo-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Next Best Offer API
  version: 0.1.0
  description: >
    REST API для управления покупателями, продуктами, событиями и получения
    Next Best Offer с пояснениями причин.
servers:
  - url: https://api.example.com
    description: Production (replace with actual host)
  - url: https://staging.api.example.com
    description: Staging
...
paths:
...
  /nbo/{customer_id}:
    get:
      tags: [Recommendations]
      summary: Получить рекомендации для покупателя
      operationId: getNextBestOffer
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - in: query
          name: channel
          description: Канал, для которого запрашивается рекомендация
          schema:
            type: string
        - in: query
          name: variant
          description: Вариант эксперимента (`control`/`treatmentA`/`treatmentB`)
          schema:
            type: string
      responses:
        '200':
          description: Рекомендации готовы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        '202':
          description: Рекомендации ещё считаются, повторите запрос позже
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationPendingResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  parameters:
    CustomerId:
      name: customer_id
      in: path
      required: true
      description: UUID покупателя
      schema:
        type: string
        format: uuid
...
  schemas:
...
    RecommendationResponse:
      type: object
      required: [status, customer_id, offers, generated_at]
      properties:
        status:
          type: string
          enum: [ready]
        customer_id:
          type: string
          format: uuid
        offers:
          type: array
          items:
            type: object
            required: [product_id, score, reason]
            properties:
              product_id:
                type: string
              score:
                type: number
                format: float
              reason:
                type: string
              variant:
                type: string
              features_snapshot:
                type: object
                additionalProperties: true
          maxItems: 10
        generated_at:
          type: string
          format: date-time
        model_version:
          type: string
        experiment_variant:
          type: string
    RecommendationPendingResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending]
        retry_after:
          type: string
          format: date-time
        reason:
          type: string
  responses:
...
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Внутренняя ошибка сервиса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Рекомендации временно недоступны
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        trace_id:
          type: string